+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed `:focus` styles and replaced with `:focus-visible`
 - Light accent blocks appear normal accent colours in dark themes
+- Added parsed meta tags to initial `gtm.js` analytics event
 
 ### Deprecated
 ### Removed